+++ include/hs_routes.h
@@ -2,7 +2,6 @@
 #define __HS_ROUTES_H__
 
 #include "hs_route.h"
-#include "hs_routes_auth.h"
 #include "hs_routes_common.h"
 #include "hs_routes_error.h"
 #include "hs_routes_favicon.h"
@@ -10,6 +9,7 @@
 #include "hs_routes_payload_limit.h"
 #include "hs_routes_powered_by.h"
 #include "hs_routes_redirection.h"
+#include "hs_routes_security.h"
 #include "hs_routes_session.h"
 #include "hs_routes_static.h"